Added middleware to handle authentication. Added hooks and pages for Login and Signup forms. Users can only access if authenticated and logged in.

+++ frontend/src/components/JobDetails.js
@@ -1,18 +1,26 @@
 import { useJobsContext } from "../hooks/useJobsContext.js";
+import { useAuthContext } from "../hooks/useAuthContext.js";
 
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const JobDetails = ({ job }) => {
     const { dispatch } = useJobsContext();
+    const { user } = useAuthContext();
+
     const { from } = job;
     const { to } = job;
     const fromAddress = `${from.street1} ${from.street2} ${from.city}, ${from.state.toUpperCase()} ${from.zipcode}`;
     const toAddress = `${to.street1} ${to.street2} ${to.city}, ${to.state.toUpperCase()} ${to.zipcode}`;
 
     const handleClick = async () => {
+        if (!user) return;
+
         const response = await fetch('http://localhost:4000/api/jobs/' + job._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
 
         const json = await response.json();
